# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
...
  response.ApiResponseMerchantMonthlyAmount:
    properties:
      data:
        description: Array of monthly amounts
        items:
          $ref: '#/definitions/response.MerchantResponseMonthlyAmount'
        type: array
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseMerchantYearlyAmount:
    properties:
      data:
        description: Array of yearly amounts
        items:
          $ref: '#/definitions/response.MerchantResponseYearlyAmount'
        type: array
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseTopupMonthAmount:
    properties:
      data:
        description: Array of monthly amount data
        items:
          $ref: '#/definitions/response.TopupMonthAmountResponse'
        type: array
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseTopupYearAmount:
    properties:
      data:
        description: Array of yearly amount data
        items:
          $ref: '#/definitions/response.TopupYearlyAmountResponse'
        type: array
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant/stats/amount/monthly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantMonthlyAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly merchant amounts
      tags:
      - Stats Merchant
...
  /api/merchant/stats/amount/yearly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantYearlyAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly merchant amounts
      tags:
      - Stats Merchant
...
  /api/topup/stats/amount/monthly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup amounts
      tags:
      - Stats Topup
  /api/topup/stats/amount/monthly/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup amounts by card number
      tags:
      - Stats Topup
  /api/topup/stats/amount/yearly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupYearAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup amounts
      tags:
      - Stats Topup
  /api/topup/stats/amount/yearly/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupYearAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup amounts by card number
      tags:
      - Stats Topup
...
  /api/topup/stats/status/monthly/failed:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup status failed stats
      tags:
      - Stats Topup
  /api/topup/stats/status/monthly/failed/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup status failed stats by card number
      tags:
      - Stats Topup
  /api/topup/stats/status/monthly/success:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup status success stats
      tags:
      - Stats Topup
  /api/topup/stats/status/monthly/success/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly topup status success stats by card number
      tags:
      - Stats Topup
  /api/topup/stats/status/yearly/failed:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup status failed stats
      tags:
      - Stats Topup
  /api/topup/stats/status/yearly/failed/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupYearStatusFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup status failed stats by card number
      tags:
      - Stats Topup
  /api/topup/stats/status/yearly/success:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup status success stats
      tags:
      - Stats Topup
  /api/topup/stats/status/yearly/success/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTopupMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly topup status success stats by card number
      tags:
      - Stats Topup
...
  /api/transaction/stats/amount/monthly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction amounts
      tags:
      - Stats Transaction
  /api/transaction/stats/amount/monthly/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction amounts by card number
      tags:
      - Stats Transaction
  /api/transaction/stats/amount/yearly:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionYearAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly transaction amounts
      tags:
      - Stats Transaction
  /api/transaction/stats/amount/yearly/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionYearAmount'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get yearly transaction amounts by card number
      tags:
      - Stats Transaction
...
  /api/transaction/stats/status/monthly/failed:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction status failed stats
      tags:
      - Stats Transaction
  /api/transaction/stats/status/monthly/failed/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthStatusFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction status failed stats by card number
      tags:
      - Stats Transaction
  /api/transaction/stats/status/monthly/success:
    get:
      consumes:
      - application/json
      parameters:
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction status success stats
      tags:
      - Stats Transaction
  /api/transaction/stats/status/monthly/success/{card_number}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Card Number
        in: path
        name: card_number
        required: true
        type: string
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      - description: Month
        in: query
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponseTransactionMonthStatusSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get monthly transaction status success stats by card number
      tags:
      - Stats Transaction
...
swagger: "2.0"